Create CAT random walks
The mean return of 0.00001 and σ of 0.00077 which we got from the data analysis are then taken and used here to generate random walks that simulate the stock. note - I made the number of steps 19800 because theirs that many seconds in a trading day. This has likely reduced the validity of the data as I'm using minute average returns on a 1-second timeframe. This is not an issue for demonstration preposes